use a RetryStrategy in server RTU mode
