+++ ffi/rodbus-schema/src/server.rs
@@ -85,9 +85,9 @@ pub(crate) fn build_server(
             "Serial port settings",
         )?
         .param(
-            "port_retry_delay",
-            DurationType::Milliseconds,
-            "How long to wait before re-opening the serial port",
+            "retry",
+            common.retry_strategy.clone(),
+            "Parameters that control How long to wait before re-opening the serial port",
         )?
         .param(
             "endpoints",
