+++ rodbus/src/serial/server.rs
@@ -1,15 +1,14 @@
 use crate::common::phys::PhysLayer;
 use crate::server::task::SessionTask;
 use crate::server::RequestHandler;
-use crate::{RequestError, SerialSettings, Shutdown};
-use std::time::Duration;
+use crate::{RequestError, RetryStrategy, SerialSettings, Shutdown};
 
 pub(crate) struct RtuServerTask<T>
 where
     T: RequestHandler,
 {
     pub(crate) port: String,
-    pub(crate) port_retry_delay: Duration,
+    pub(crate) retry: Box<dyn RetryStrategy>,
     pub(crate) settings: SerialSettings,
     pub(crate) session: SessionTask<T>,
 }
@@ -22,6 +21,7 @@ where
         loop {
             match crate::serial::open(&self.port, self.settings) {
                 Ok(serial) => {
+                    self.retry.reset();
                     tracing::info!("opened port");
                     // run an open port until shutdown or failure
                     let mut phys = PhysLayer::new_serial(serial);
@@ -29,18 +29,20 @@ where
                         return Shutdown;
                     }
                     // we wait here to prevent any kind of rapid retry scenario if the port opens and immediately fails
-                    tracing::warn!("waiting {:?} to reopen port", self.port_retry_delay);
-                    if let Err(Shutdown) = self.session.sleep_for(self.port_retry_delay).await {
+                    let delay = self.retry.after_disconnect();
+                    tracing::warn!("waiting {:?} to reopen port", delay);
+                    if let Err(Shutdown) = self.session.sleep_for(delay).await {
                         return Shutdown;
                     }
                 }
                 Err(err) => {
+                    let delay = self.retry.after_failed_connect();
                     tracing::warn!(
                         "unable to open serial port, retrying in {:?} - error: {}",
-                        self.port_retry_delay,
+                        delay,
                         err
                     );
-                    if let Err(Shutdown) = self.session.sleep_for(self.port_retry_delay).await {
+                    if let Err(Shutdown) = self.session.sleep_for(delay).await {
                         return Shutdown;
                     }
                 }
