+++ rodbus/src/server/mod.rs
@@ -19,6 +19,7 @@ pub(crate) const SERVER_SETTING_CHANNEL_CAPACITY: usize = 8;
 
 use crate::error::Shutdown;
 
+use crate::RetryStrategy;
 pub use address_filter::*;
 pub use handler::*;
 pub use types::*;
@@ -96,7 +97,7 @@ pub async fn spawn_tcp_server_task<T: RequestHandler>(
 ///
 /// * `path` - Path to the serial device. Generally `/dev/tty0` on Linux and `COM1` on Windows.
 /// * `settings` - Serial port settings
-/// * `port_retry_delay` - How often the task will attempt to open the serial port if it fails.
+/// * `retry` - A boxed trait object that controls when opening the serial port is retried after a failure
 /// * `handlers` - A map of handlers keyed by a unit id
 /// * `decode` - Decode log level
 ///
@@ -105,7 +106,7 @@ pub async fn spawn_tcp_server_task<T: RequestHandler>(
 pub fn spawn_rtu_server_task<T: RequestHandler>(
     path: &str,
     settings: crate::serial::SerialSettings,
-    port_retry_delay: std::time::Duration,
+    retry: Box<dyn RetryStrategy>,
     handlers: ServerHandlerMap<T>,
     decode: DecodeLevel,
 ) -> Result<ServerHandle, std::io::Error> {
@@ -121,7 +122,7 @@ pub fn spawn_rtu_server_task<T: RequestHandler>(
 
     let mut rtu = crate::serial::server::RtuServerTask {
         port: path.to_string(),
-        port_retry_delay,
+        retry,
         settings,
         session,
     };
